Destry wants to move this to the wiki.
Let him have what he wants. I'm not sure how he plans
to communicate with people, but at least this forces
people to actually update the wiki rather than announcing
that something is wrong. It's a W-I-K-I. If The idea behind
wikis is that EVERYONE contributes, giving it wider
coverage and multiple viewpoints and ideas.

I'm not sure if Destry is aware that users can not freely
register on the wiki, and this forum was the place to
ask an account; unless you really want to personally
bother someone, who's email address you will not even
know.