Quote includes all quotable paragraphs.
Move the link to the heading, support Textile syntax
characters in account names. Due to this it requires
escape syntax which requires wrapping in spaces and
thus we must change the strong to a heading.

If the quote doesn't have a message, only adds a back
link to the original message.