Fix the navigation loading on desktop.
The navigation must either:

* Be desktop first.
* Expect that JavaScript is always supported.
* Or not have collapsing.

The mobile first layout causes the collapsing animation
run on each page load on desktop devices when it
should not.

The menu starts opened (due to .js rules), then collapses,
and opens again once the menu script has been executed.

When changed to desktop fist it still does jagged loading
due to the .js class, but now it only happens on mobiles
rather than on both.

To completely fix the loading issues on both platforms,
the .js class must be removed.