SFS bot no longer bans, but just blocks.
We now have CAPTCHA which limits request activity.
Its better be safe than randomly ban IPs. Instead, show
injected alerts and limit request activity by sleeping.

Even if we ban, you can still reach the 20000 daily
request limit SFS has, especially if your IP isn't flagged.

The main reason is FluxBB and the fact that the bans
do not scale. The speed we are going, we are going to
run out of memory.